requirements.txt
openai>=1.59.0
streamlit>=1.41.0
python-dotenv>=1.0.1
pytest>=8.3.0
pytest-asyncio>=0.25.0
